DX-1622: Rename diagram.rb to plantuml_diagram.rb
