+++ lib/kramdown-plantuml/plantuml_diagram.rb
