+++ lib/kramdown_html.rb
@@ -5,7 +5,7 @@ require 'kramdown-parser-gfm'
 require_relative 'kramdown-plantuml/log_wrapper'
 require_relative 'kramdown-plantuml/plantuml_error'
 require_relative 'kramdown-plantuml/options'
-require_relative 'kramdown-plantuml/diagram'
+require_relative 'kramdown-plantuml/plantuml_diagram'
 require_relative 'kramdown-plantuml/jekyll_provider'
 
 module Kramdown
